+++ app/controllers/sessions_controller.rb
@@ -1,8 +1,8 @@
 require 'digest'
 require 'jwt'
+require 'user_helper'
 
 class SessionsController < ApplicationController
-
   def create
     @user = User.find_by(name: sessions_params[:name])
     if @user&.authenticate(sessions_params[:password])
@@ -16,18 +16,6 @@ class SessionsController < ApplicationController
     params.require(:session).permit(:name, :password)
   end
 
-  def authenticate_user
-    token = request.headers['token']
-    begin
-      payload = JWT.decode(token, secret_key, false)
-      @user = User.find(JSON.parse(payload[0])['id'])
-    rescue JWT::DecodeError => e
-      render json: { error: e.message }, status: :unauthorized
-    rescue ActiveRecord::RecordNotFound => e
-      render json: { error: e.message }, status: :unauthorized
-    end
-  end
-
   def encode_token
     payload = { 'id' => @user[:id], 'created_at' => @user[:created_at] }
     JWT.encode(payload.to_json, secret_key, 'none')
@@ -36,6 +24,6 @@ class SessionsController < ApplicationController
   private
 
   def secret_key
-    'kisw@hili7!'
+    'Kiswahili@7!'
   end
 end
